Include more order restriction code
Includes codes that only have one of the three conditions in the order restriction model to see if results where M0> M2 is due to combining the three conditions (HT-HS-, HT+HS-, HT-HS+) into one level in the predictor. 

Found that in all 3 order restriction models M0  still > M2 suggesting it is not due combining the levels in the condition predictor. 

also modified the code for M1 to adjust prior to account addition increase in possible ways where the model can be correct