Includes codes to magnify parameter effect by 100
this is done for a sanity check to make sure that it is not something fundamentally wrong with the code. However, due to the specific implementation means that modified parameters are largely in effective in changing BFs. For example M3 the number of usable cases is depended on Identify =< Imagine =< HTPHSP so magnifying HTPHSP does not have much an change on BF since Identify =< Imagine = HTPHSP is pretty much set regardless value of HTPHSP

I think the disparity between results of parameter estimation and model comparison is due to model complexity rather than something fundamentally wrong with the implementation of the code.